% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{is_deal}
\alias{is_deal}
\title{Check if the object is a valid deal}
\usage{
is_deal(a)
}
\arguments{
\item{a}{list}
}
\value{
logical value for whether a list is a deal
}
\description{
Check if the object is a valid deal
}
\examples{
a <- list(
  data.frame(
    x=c(2500, 0),
    p=c(0.33, 0.67)),
  data.frame(
    x=c(2400, 0),
    p=c(0.34, 0.66))
)

is_deal(a)

}
